refactored with rename. Only manual required change was for the pyproject.toml ruff.lint.per-file-ignores
