+++ app/scraping_functions/population.py
@@ -1,5 +1,5 @@
-from src.scraping_functions.population_dictionary import pop_dictionary
-from src.scraping_functions.state_abbreviations import state_abbr
+from app.scraping_functions.population_dictionary import pop_dictionary
+from app.scraping_functions.state_abbreviations import state_abbr
 
 # USE THIS FUNCTION to get a city's population. Enter in string format the city name, and the state (abbreviated or
 # full works)
