+++ app/templates/schema_diagram.html
